+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.5.0] - 2020-01-22
+
 ### Added
 - Add Signal/PSR-14 event to decide about embedding of markup (#29)
 - Update schema definition to version 6.0 (#3)
@@ -142,7 +144,8 @@ Initial release
 - View helpers for usage in Fluid templates
 
 
-[Unreleased]: https://github.com/brotkrueml/schema/compare/v1.4.2...HEAD
+[Unreleased]: https://github.com/brotkrueml/schema/compare/v1.5.0...HEAD
+[1.5.0]: https://github.com/brotkrueml/schema/compare/v1.4.2...v1.5.0
 [1.4.2]: https://github.com/brotkrueml/schema/compare/v1.4.1...v1.4.2
 [1.4.1]: https://github.com/brotkrueml/schema/compare/v1.4.0...v1.4.1
 [1.4.0]: https://github.com/brotkrueml/schema/compare/v1.3.1...v1.4.0